+++ sdk/@launchdarkly/observability-react-native/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.7.1](https://github.com/launchdarkly/observability-sdk/compare/observability-react-native-0.7.0...observability-react-native-0.7.1) (2026-03-05)
+
+
+### Bug Fixes
+
+* ldclient dependencies ([#407](https://github.com/launchdarkly/observability-sdk/issues/407)) ([65a5e6a](https://github.com/launchdarkly/observability-sdk/commit/65a5e6a1999c9e66c7f4011f512d17de256f919c))
+
 ## [0.7.0](https://github.com/launchdarkly/observability-sdk/compare/observability-react-native-0.6.0...observability-react-native-0.7.0) (2025-10-15)